shopify_agent/requirements.txt
langchain
langchain-core
django
dotenv
langchain-community
langchain-google-genai
pymupdf
langchain-huggingface
google-genai
tabulate